+++ integration_tests/open_R13.py
@@ -1,14 +1,14 @@
 #!/usr/bin/env python3
 #coding:utf-8
 # Author:  mozman -- <mozman@gmx.at>
-# Purpose: open a DXF R14 file without proper setup layouts
+# Purpose: open a DXF R13 file without plot style name structure setup
 # Created: 05.03.2016
 # Copyright (C) 2016, Manfred Moitzi
 # License: MIT License
 
 import ezdxf
 
-FILENAME = "small_r14.dxf"
+FILENAME = "small_r13.dxf"
 
 
 def main():
@@ -26,7 +26,7 @@ def main():
     msp = dwg.modelspace()
     msp.add_line((0, 0), (10, 3))
 
-    dwg.saveas("small_r14_converted_AC1015.dxf")
+    dwg.saveas("small_r13_converted_AC1015.dxf")
 
 
 if __name__ == '__main__':